move node files back
