--- src/node/jacdac-node.ts
@@ -1,4 +1,0 @@
-export * from "./nodesocket"
-export * from "./nodespi"
-export * from "./nodewebserialio"
-export * from "./nodewebusb"
